[hooks] Fix check-shader-includes hook (#1551)
* Prevent false positive from being reported

* Add standalone version of check_shader_includes

* Remove old shell scripts

* Add support for exclusions
