--- Tools/git-hook/pre-commit
@@ -1,14 +1,0 @@
-#!/bin/sh
-
-cd "$(dirname "$0")/precommit"
-
-for hook in *.sh; do
-	sh $hook # On windows this will be executed using git's integrated bash (automatic)
-	RESULT=$?
-	if [ $RESULT != 0 ]; then
-		echo "precommit/$hook returned non-zero: $RESULT, abort commit"
-		exit $RESULT
-	fi
-done
-
-exit 0
