--- Tools/git-hook/pre-push
@@ -1,14 +1,0 @@
-#!/bin/sh
-
-cd "$(dirname "$0")/prepush"
-
-for hook in *.sh; do
-	sh $hook # On windows this will be executed using git's integrated bash (automatic)
-	RESULT=$?
-	if [ $RESULT != 0 ]; then
-		echo "prepush/$hook returned non-zero: $RESULT, abort push"
-		exit $RESULT
-	fi
-done
-
-exit 0
